Added object parser functions for determining which tables need to be recursively visited, split SQLStringConcationation into 4 parts.
