+++ src/main/java/hiof/gruppe1/Estivate/objectParsers/IObjectParser.java
@@ -2,10 +2,11 @@ package hiof.gruppe1.Estivate.objectParsers;
 
 import hiof.gruppe1.Estivate.Objects.SQLAttribute;
 
+import java.util.ArrayList;
 import java.util.HashMap;
 
 public interface IObjectParser {
     public <T> HashMap<String, SQLAttribute> parseObjectToAttributeList(Object object);
-
+    public <T> ArrayList<String> getSubElementList(T castingClass);
     public <T> T parseAttributeListToObject(Class<T> castTo, HashMap<String, SQLAttribute> attributeList);
 }
